requirements.txt
Flask==2.2.5
numpy==1.24.3
ipython==8.12.0
Pillow==9.5.0
transformers==4.30.2
ultralytics==8.0.151
